# openapi.yaml
openapi: 3.0.0
info:
  title: SOGo6 Groupware Server API
  version: 6.0.0-alpha1
  description: SOGo6 is a modern groupware server providing email, calendar, contacts,
    tasks, and collaboration features. This REST API allows clients to interact with
    all SOGo6 functionality.
  contact:
    name: SOGo6 Team
    url: https://github.com/Alinto/sogo6
    email: sogo@alinto.com
  license:
    name: GNU General Public License v3.0 or later
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
- url: https://api.sogo.example.com/api/v1
  description: Production
- url: http://localhost:5000/api/v1
  description: Development
- url: /api/v1
  description: Relative path
paths:
...
  /api/admin/v1/admin/users:
    get:
      tags:
      - admin
      summary: List users
      description: List users
      operationId: admin_get__api_admin_v1_admin_users
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          description: Internal server error
    post:
      tags:
      - admin
      summary: Create user
      description: Create user
      operationId: admin_post__api_admin_v1_admin_users
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          description: Internal server error
  /api/admin/v1/admin/users/<user_id>:
    get:
      tags:
      - admin
      summary: Get user details
      description: Get user details
      operationId: admin_get__api_admin_v1_admin_users_user_id
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          description: Internal server error
    put:
      tags:
      - admin
      summary: Update user
      description: Update user
      operationId: admin_put__api_admin_v1_admin_users_user_id
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          description: Internal server error
    delete:
      tags:
      - admin
      summary: Delete user
      description: Delete user
      operationId: admin_delete__api_admin_v1_admin_users_user_id
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          description: Internal server error
...
components:
  schemas:
    Error:
      type: object
      properties:
        error_code:
          type: string
          example: S000000
        error_msg:
          type: string
          example: No Error
        data:
          type: object
          nullable: true
      required:
      - error_code
      - error_msg
    Success:
      type: object
      properties:
        data:
          type: object
        error_code:
          type: string
          example: S000000
        error_msg:
          type: string
          example: No Error
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /api/user/v1/auth/login
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Too many requests
      headers:
        X-RateLimit-Retry-After:
          schema:
            type: integer
security:
- bearerAuth: []
...